Clean up code: mode weatherbit methods and exchangeRate methods to ./routes
